+++ instant/src/main/java/com/arnyminerz/escalaralcoiaicomtat/ui/viewmodel/AreasViewModel.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.liveData
 import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.area.Area
 import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.area.loadAreas
 import com.arnyminerz.escalaralcoiaicomtat.core.shared.AREAS
+import com.arnyminerz.escalaralcoiaicomtat.core.ui.viewmodel.DataClassViewModel
 import timber.log.Timber
 import kotlin.coroutines.resume
 import kotlin.coroutines.suspendCoroutine
